feat: Add export function splitted in 25 lines, Norm fixes in main.
