+++ include/minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pmarkaid <pmarkaid@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: dbejar-s <dbejar-s@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/01 15:11:45 by pmarkaid          #+#    #+#             */
-/*   Updated: 2024/08/29 13:30:52 by pmarkaid         ###   ########.fr       */
+/*   Updated: 2024/08/29 22:12:13 by dbejar-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -222,6 +222,17 @@ int					change_directory(char *path, char *home);
 void				check_save_env(char *var, t_macro *macro, int size);
 void				update_environment(t_macro *macro, char *oldpwd,
 						char *path);
+void				print_export_var(char *var, char *value);
+void				export_argless(t_macro *macro);
+int					check_export(char *arg);
+char				*remove_quotes(char *str);
+t_macro				*init_macro(char **envp, char **argv);
+char				*grab_home(t_macro *macro);
+t_macro				*start_env(t_macro *macro, char **argv);
+int					in_root(char *path);
+int					in_home(t_macro *macro);
+char				*upper_than_home(t_macro *macro);
+char				*create_path(t_macro *macro);
 
 /* error */
 int					error_msg(char *msg, int exit_code);
