+++ src/error.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   error.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pmarkaid <pmarkaid@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: dbejar-s <dbejar-s@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/10 13:00:20 by pmarkaid          #+#    #+#             */
-/*   Updated: 2024/08/29 15:06:15 by pmarkaid         ###   ########.fr       */
+/*   Updated: 2024/08/29 22:13:31 by dbejar-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int	error_msg(char *file, int exit_code)
 	return (exit_code);
 }
 
-void exit_error(char *file, char *msg, t_macro *macro, int exit_code)
+void	exit_error(char *file, char *msg, t_macro *macro, int exit_code)
 {
 	ft_putstr_fd("minishell: ", 2);
 	ft_putstr_fd(file, 2);
