+++ src/validation.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   validation.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pmarkaid <pmarkaid@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: dbejar-s <dbejar-s@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/08 22:35:57 by pmarkaid          #+#    #+#             */
-/*   Updated: 2024/08/29 15:10:56 by pmarkaid         ###   ########.fr       */
+/*   Updated: 2024/08/29 22:41:12 by dbejar-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -82,13 +82,12 @@ void	search_executable(t_macro *macro, t_cmd *cmd)
 	}
 }
 
-void validation(t_macro *macro, t_cmd *cmd)
+void	validation(t_macro *macro, t_cmd *cmd)
 {
-	
 	if (ft_strchr("./", cmd->cmd_arg->value[0]) == NULL)
 		search_executable(macro, cmd);
 	validate_access(cmd->cmd_arg->value, macro);
-	if(validate_redirections(cmd->redir, macro) == -1)
+	if (validate_redirections(cmd->redir, macro) == -1)
 	{
 		free_ins(macro);
 		exit(1);
